requirements-npu.txt

torch_npu==2.2.0
torchvision==2.2.0
torchaudio==0.17.0

modelscope==1.33.0
funasr==1.2.6
addict==2.4.0
datasets==2.21.0

# 核心依赖
# 华为CANN很多算子和旧版transformers对Numpy2.0支持极差
numpy==1.26.4

transformers==4.29.2

# 音频处理依赖
av==14.4.0
soundfile
librosa
pydub
pyannote.audio
ffmpeg-python

# --- Web 服务 ---
fastapi
uvicorn
python-multipart
requests
aiofiles
websockets==12.0
simplejson==3.20.1

# 普通话识别
openai-whisper==20250625

# --- ⚠️ 性能警告区 ---
# ctranslate2 和 faster-whisper 不支持华为 NPU
# 它们会以纯 CPU 模式运行，推理速度会比 NPU 慢几十倍
# 建议保留仅作对比或备用，生产环境请尽量移除
ctranslate2==4.4.0

# faster-whisper ctranslate2 和 faster-whisper 不支持华为 NPU